// backend/src/db/schemas/infra-files.ts
// Code generated by automation script, DO NOT EDIT.
// Automated by pulling database and generating zod schema
// To update. Just run npm run generate:schema
// Written by akhilmhdh.

import { z } from "zod";

import { TImmutableDBKeys } from "./models";

export const InfraFilesSchema = z.object({
  id: z.string().uuid(),
  projectId: z.string(),
  name: z.string(),
  content: z.string().default(""),
  createdAt: z.date(),
  updatedAt: z.date()
});

export type TInfraFiles = z.infer<typeof InfraFilesSchema>;
export type TInfraFilesInsert = Omit<z.input<typeof InfraFilesSchema>, TImmutableDBKeys>;
export type TInfraFilesUpdate = Partial<Omit<z.input<typeof InfraFilesSchema>, TImmutableDBKeys>>;